requirements.txt
#pip install -r requirements.txt
#pip install --upgrade pip
requests
python-slugify
glob2
pyyaml
Jinja2
GitPython